requirements.txt
# 浏览器自动化
playwright>=1.40.0

# 图像处理与 PDF 生成
Pillow>=10.0.0
img2pdf>=0.5.0
reportlab>=4.0.0

# 配置与 CLI
pyyaml>=6.0
click>=8.0.0

# 工具
tqdm>=4.66.0
